requirements.txt
fastapi
qrcode
aiofiles
python-jwt
pillow